[BE] refactor: Service-ServiceImpl 패턴 적용 & 중복 파일 삭제
