+++ backend/src/main/java/com/proj/withus/controller/RoomController.java
@@ -3,9 +3,8 @@ package com.proj.withus.controller;
 import com.proj.withus.domain.Room;
 import com.proj.withus.domain.dto.CreateRoomReq;
 import com.proj.withus.domain.dto.EnterRoomRes;
-import com.proj.withus.domain.dto.ModifyNicknameReq;
 import com.proj.withus.domain.dto.ModifyRoomReq;
-import com.proj.withus.service.RoomService;
+import com.proj.withus.service.RoomServiceImpl;
 import com.proj.withus.util.JwtUtil;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -21,7 +20,7 @@ import java.util.Optional;
 @RequestMapping("/api/rooms")
 public class RoomController {
 
-    private final RoomService roomService;
+    private final RoomServiceImpl roomServiceImpl;
     private final JwtUtil jwtUtil;
 
     @PostMapping
@@ -37,7 +36,7 @@ public class RoomController {
 //        } catch (Exception e) {
 //            return new ResponseEntity<String>("권한이 없는 유저입니다.", HttpStatus.UNAUTHORIZED);
 //        }
-        Room newRoom = roomService.createRoom(createRoomReq);
+        Room newRoom = roomServiceImpl.createRoom(createRoomReq);
         return new ResponseEntity<Room>(newRoom, HttpStatus.OK);
     }
 
@@ -52,7 +51,7 @@ public class RoomController {
             return new ResponseEntity<String>("권한이 없는 유저입니다.", HttpStatus.UNAUTHORIZED);
         }
 
-        Optional<Room> room = roomService.enterRoom(roomId, memberId);
+        Optional<Room> room = roomServiceImpl.enterRoom(roomId, memberId);
         // Optional<>은 isPresent()로 확인!
         if (!room.isPresent()) {
             return new ResponseEntity<String>("존재하지 않는 방입니다.", HttpStatus.BAD_REQUEST);
@@ -61,8 +60,8 @@ public class RoomController {
             enterRoomRes.setRoomId(room.get().getId());
             enterRoomRes.setRoomType(room.get().getType());
             enterRoomRes.setCode(String.valueOf(room.get().getCode()));
-            enterRoomRes.setHostId(roomService.getHostId(roomId));
-            enterRoomRes.setPlayers(roomService.getPlayerList(roomId)); // List<>를 이렇게 set하는게 맞나..
+            enterRoomRes.setHostId(roomServiceImpl.getHostId(roomId));
+            enterRoomRes.setPlayers(roomServiceImpl.getPlayerList(roomId)); // List<>를 이렇게 set하는게 맞나..
             return new ResponseEntity<EnterRoomRes>(enterRoomRes, HttpStatus.OK);
         }
 
@@ -76,7 +75,7 @@ public class RoomController {
 
         Long memberId = jwtUtil.extractMemberId(token); // try-catch 뺌
         try {
-            roomService.leaveRoom(roomId, memberId);
+            roomServiceImpl.leaveRoom(roomId, memberId);
         } catch (Exception e) {
             return new ResponseEntity<String>("오류가 발생했습니다.", HttpStatus.BAD_REQUEST);
         }
@@ -94,7 +93,7 @@ public class RoomController {
         try {
             Long id = jwtUtil.extractMemberId(token);
             // 방장 체크 // 이렇게 깊은건 어떻게 처리하는게 깔끔한지 알아보기 (depth 3 이상)
-            Long hostId = roomService.getHostId(roomId);
+            Long hostId = roomServiceImpl.getHostId(roomId);
             if (hostId != id) {
                 return new ResponseEntity<String>("방장이 아닙니다.", HttpStatus.FORBIDDEN);
             }
@@ -106,7 +105,7 @@ public class RoomController {
         if (!isValid) {
             return new ResponseEntity<String>("토큰이 만료되었습니다.", HttpStatus.UNAUTHORIZED);
         }
-        roomService.modifyRoom(modifyRoomReq, roomId); // 반환 받아도 됨
+        roomServiceImpl.modifyRoom(modifyRoomReq, roomId); // 반환 받아도 됨
         return new ResponseEntity<String>("성공", HttpStatus.OK);
     }
 
@@ -126,7 +125,7 @@ public class RoomController {
         System.out.println("ididididi: " + id); //
 
         try {
-            int isSuccess = roomService.modifyNickname(id, nickname);
+            int isSuccess = roomServiceImpl.modifyNickname(id, nickname);
             if (isSuccess != 1) {
                 System.out.println("ididididi: " + id); //
                 return new ResponseEntity<String>("권한이 없습니다.", HttpStatus.UNAUTHORIZED);
