--- backend/src/main/java/com/proj/withus/domain/dto/GoogleUserInfo.java
@@ -1,17 +1,0 @@
-//package com.proj.withus.domain.dto;
-//
-//import lombok.*;
-//
-//@Data
-//@NoArgsConstructor
-//@AllArgsConstructor
-//@ToString
-//@Builder
-//@Getter
-//public class GoogleUserInfo {
-//
-//    private Long id;
-//    private String nickname;
-//    private String email;
-//    private String profile;
-//}
