--- backend/src/main/java/com/proj/withus/service/KakaoService.java
@@ -1,4 +1,0 @@
-//package com.proj.withus.service;
-//
-//public interface KakaoService {
-//}
