+++ backend/src/main/java/com/proj/withus/service/MemberService.java
@@ -1,8 +1,11 @@
-//package com.proj.withus.service;
-//
-//import com.proj.withus.domain.dto.GoogleUserInfo;
-//
-//public interface MemberService {
-//
-//    public void saveGoogle(GoogleUserInfo userInfo, String accessToken);
-//}
+package com.proj.withus.service;
+
+import com.proj.withus.domain.Member;
+
+public interface MemberService {
+
+    public Member getMemberInfo(Long id);
+    public Member updateMember(Long id, String nickname);
+    public Member deleteMember(Long id);
+
+}
