--- backend/src/main/java/com/proj/withus/service/OauthService.java
@@ -1,4 +1,0 @@
-//package com.proj.withus.service;
-//
-//public interface OauthService {
-//}
